# documentation/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Shonet API",
        "contact": {
            "name": "The Shonet",
            "url": "https://www.theshonet.com/home"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/articles/{id}": {
          	"get": {
            		"tags": [
              		"Articles"
            		],
            		"summary": "Get Journal and Product article",
	            	"description": "Get specific Journal or Product article for editing article",
	            	"operationId": "articles.view",
	            	"produces": [
	              	"application/json"
	            	],
	            	"parameters": [
              		{
                			"name": "token",
                			"in": "header",
                			"description": "authority token",
                			"required": true,
                			"type": "string"
              		},
              		{
                			"in": "path",
                			"name": "id",
                			"required": true,
                			"type": "integer",
                			"minimum": 1,
                			"description": "search article by id parameter"
              		}
            		],
            		"responses": {
              		"400": {
                			"description": "Bad Request, required form data"
              		},
              		"401": {
                			"description": "Not Authorized, api key not valid"
              		},
              		"403": {
                			"description": "Forbidden Access, You not have priviledges to access this API"
              		},
              		"404": {
                			"description": "Articles Not Found"
              		},
              		"500": {
                			"description": "Internal Server error"
              		},
              		"200": {
                			"description": "Success get article"
              		}
            		}
          	},
          	"put": {
            		"tags": [
            	  		"Articles"
            		],
            		"summary": "Update Journal or Product article",
            		"description": "Update Journal or Product article with all required body",
            		"operationId": "articles.update",
            		"produces": [
              		"application/json"
            		],
            		"parameters": [
              		{
                			"name": "token",
                			"in": "header",
                			"required": true,
                			"description": "authority token",
                			"type": "string"
              		},
              		{
                			"name": "id",
                			"in": "path",
                			"required": true,
                			"description": "add id to specific update article",
                			"type": "integer"
              		},
              		{
                			"name": "body",
                			"in": "body",
                			"description": "object that need for update article",
                			"required": true,
                			"schema": {
                  			"type": "object",
                  			"required":[
                    				"title",
                    				"content",
                    				"image",
                    				"seo_keyword",
                    				"type",
                    				"status",
                    				"request_publish_date",
                    				"category",
                    				"tags",
                    				"products"
                  			],
                  			"properties": {
                    				"title": {
                      				"type": "string",
                      				"format": "text",
                      				"example": "Journal article edit new"
                    				},
                    				"content": {
                      				"type": "string",
                      				"format": "text",
                      				"example": "<p>Can you input HTML code as string</p>"
                    				},
                    				"image": {
                      				"type": "string",
                      				"format": "byte",
                      				"example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAASABIAAD/4QBMRXhpZgAATU0AKgAAAAgAAgESAAMAAAABAAEAAIdpAAQAAAABAAAAJgAAAAAAAqACAAQAAAABAAACVKADAAQAAAABAAACSgAAAAD/4QkhaHR0cDovL25zLmFkb2JlLmNvbS94YXAvMS4wLwA8P3hwYWNrZXQgYmVnaW49Iu+7vyIgaWQ9Ilc1TTBNcENlaGlIenJlU3pOVGN6a2M5ZCI/PiA8eDp4bXBtZXRhIHhtbG5zOng9ImFkb2JlOm5zOm1ldGEvIiB4OnhtcHRrPSJYTVAgQ29yZSA1LjQuMCI+IDxyZGY6UkRGIHhtbG5zOnJkZj0iaHR0cDovL3d3dy53My5vcmcvMTk5OS8wMi8yMi1yZGYtc3ludGF4LW5zIyI/mab6en6v+v+GNy3d2loJiotbcKqRhQII8KNoOABHgDJJwO/NID//Z"
                    				},
                    				"type": {
                      				"type": "string",
                      				"enum": ["P", "J"],
                      				"example": "P"
                    				},
                    				"status": {
                      				"type": "string",
                      				"enum": ["D", "S", "P"],
                      				"example": "D"
                    				},
                    				"request_publish_date": {
                      				"type": "string",
                      				"format": "date-time"
                    				},
                    				"category": {
                      				"type": "array",
                      				"items": [
                        					{
                          					"type": "object",
                          					"properties": {
                            						"id": {
                              						"type": "integer",
                              						"format": "int64",
                              						"example": 6
                            						}
                          					}
                        					},
                        					{
                          					"type": "object",
                          					"properties": {
                            						"id": {
                              						"type": "integer",
                              						"format": "int64",
                              						"example": 35
                            						}
                          					}
                        					}
                      				]
                    				},
                    				"tags": {
                      				"type": "array",
                      				"items": [
                        					{
                          					"type": "object",
                          					"properties": {
                            						"id": {
                            	  						"type": "string",
                              						"format": "text",
                              						"example": "shonet"
                            						}
                          					}
                        					},
                        					{
                          					"type": "object",
                          					"properties": {
                            						"id": {
                              						"type": "string",
                              						"format": "text",
                              						"example": "cakep"
                            						}
                          					}
                        					}
                      				]
                    				},
                    				"products": {
                      				"type": "array",
                      				"items": [
                        					{
                          					"type": "object",
                          					"properties": {
                            						"id": {
                              						"type": "integer",
                              						"format": "uint",
                              						"example": 1531
                            						}
                          					}
                        					},
                        					{
                          					"type": "object",
                          					"properties": {
                            						"id": {
                              						"type": "integer",
                              						"format": "uint",
                              						"example": 1533
                            						}
                          					}
                        					},
                        					{
                          					"type": "object",
                          					"properties": {
                            						"id": {
                              						"type": "integer",
                              						"format": "uint",
                              						"example": 1536
                            						}
                          					}
                        					}
                      				]
                    				}
                  			}
                			}
              		}
            		],
            		"responses": {
              		"400": {
                			"description": "Bad Request, required form data"
              		},
              		"401": {
                			"description": "Not Authorized, api key not valid"
              		},
              		"403": {
                			"description": "Forbidden Access, You not have priviledges to access this API"
              		},
              		"404": {
                			"description": "Articles Not Found"
              		},
              		"500": {
                			"description": "Internal Server error"
              		},
              		"200": {
                			"description": "Success to update articles"
              		}
            		}
          	},
          	"delete": {
            		"tags": [
              		"Articles"
            		],
            		"summary": "Deleting Journal and attach all article relations",
            		"description": "Deleting Journal and attach all article relations, with id supply",
            		"operationId": "articles.destroy",
            		"produces": [
              		"application/json"
            		],
            		"parameters": [
              		{
                			"name": "token",
                			"in": "header",
                			"required": true,
                			"description": "authority token",
                			"type": "string"
              		},
              		{
                			"name": "id",
                			"in": "path",
                			"required": true,
                			"description": "add id to specific article delete",
                			"type": "integer"
              		}
            		],
            		"responses": {
              		"400": {
                			"description": "Bad Request, required form data"
              		},
              		"401": {
                			"description": "Not Authorized, api key not valid"
              		},
              		"403": {
                			"description": "Forbidden Access, You not have priviledges to access this API"
              		},
              		"404": {
                			"description": "Articles Not Found"
              		},
              		"500": {
                			"description": "Internal Server error"
              		},
              		"204": {
                			"description": "Success to delete articles"
              		}
            		}
          	}
        },
...
	}
}